New loader interfaces (#486)
Separate out the XVIZLoaderInterface into a functional breakdown.
- LoaderInterface: Interface necessary for HOC state update using `connect()` 
- PlayableLoaderInterface: Interface supporting playback controls (ie play/pause/seek)
- XVIZLoaderInterface: Full XVIZ Log specific functionality

Update docs and expose `connect()` as a more general name than `connectToLog()`
